requirements.txt
polars
pillow